Revert "product category left sidebar and service section"
