--- app/components/productCategory/CategorySectionTitle.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const CategorySectionTitle = ({ title }) => {
-    return (
-        <div className="category-left-section-title">
-            <h2>{title}</h2>
-        </div>
-    )
-}
-
-export default CategorySectionTitle
