--- app/components/productCategory/SubCategoryListItems.jsx
@@ -1,12 +1,0 @@
-import Link from "next/link"
-
-const SubCategoryListItems = ({ categoryItem }) => {
-    const { title, path } = categoryItem;
-    return (
-        <>
-            <li><Link href={path}>{title} </Link></li>
-        </>
-    )
-}
-
-export default SubCategoryListItems
