--- app/components/productCategory/ViewMoreLink.jsx
@@ -1,11 +1,0 @@
-import Link from "next/link"
-
-const ViewMoreLink = ({ title = "View More", path = "#" }) => {
-    return (
-        <div className="view-more-link-area">
-            <Link href={path}><span>+</span> {title} </Link>
-        </div>
-    )
-}
-
-export default ViewMoreLink
