+++ public/css/common.css
@@ -30,9 +30,6 @@
     --color-949495: #949495;
     --color-ACACAC: #ACACAC;
     --color-8F8F8F: #8F8F8F;
-    --color-FFEEEE: #FFEEEE;
-    --color-646464: #646464;
-    --color-FC7024: #FC7024;
 }
 
 ul {
