+++ public/css/header.css
@@ -108,10 +108,6 @@ header {
     padding: 0 20px;
 }
 
-.observerable-header-search-area .header-search-location {
-    flex: 1.4;
-}
-
 .header-search-field {
     flex: 3;
 }
